refactor: remove leftover code
