+++ library/src/helpers/index.ts
@@ -1,9 +1,7 @@
-export * from './beautifier';
 export * from './bemClasses';
 export * from './createNestedClassName';
 export * from './extractHashData';
 export * from './scrollIntoViewOfAnchor';
-export * from './searchForNestedObject';
 export * from './generateExampleSchema';
 export * from './inContainer';
 export * from './stateHelpers';
