+++ library/src/helpers/parser.ts
@@ -6,33 +6,17 @@ import { ErrorObject, ParserReturn, FetchingSchemaInterface } from '../types';
 
 import { VALIDATION_ERRORS_TYPE } from '../constants';
 
-type ParseDocument = (content: string | any, options?: any) => Promise<any>;
-
-type ParseDocumentFromURL = (
-  url: string,
-  requestOptions?: RequestInit,
-  options?: any,
-) => Promise<any>;
-
 registerSchemaParser(openapiSchemaParser);
 registerSchemaParser(avroSchemaParser);
 
-export class Parser {
-  private parseSchema: ParseDocument;
-  private parseSchemaFromURL: ParseDocumentFromURL;
-
-  constructor() {
-    this.parseSchema = parse;
-    this.parseSchemaFromURL = parseFromUrl;
-  }
-
+class Parser {
   async parse(
     content: string | any,
     parserOptions?: any,
   ): Promise<ParserReturn> {
     try {
-      const data = await this.parseSchema(content, parserOptions);
-      return this.extractDocument(data);
+      const asyncapi = await parse(content, parserOptions);
+      return { asyncapi };
     } catch (err) {
       return this.handleError(err);
     }
@@ -43,12 +27,12 @@ export class Parser {
     parserOptions?: any,
   ): Promise<ParserReturn> {
     try {
-      const data = await this.parseSchemaFromURL(
+      const asyncapi = await parseFromUrl(
         arg.url,
         arg.requestOptions,
         parserOptions,
       );
-      return this.extractDocument(data);
+      return { asyncapi };
     } catch (err) {
       return this.handleError(err);
     }
@@ -57,31 +41,11 @@ export class Parser {
   private handleError = (err: ErrorObject): ParserReturn => {
     if (err.type === VALIDATION_ERRORS_TYPE) {
       return {
-        data: err.parsedJSON || null,
-        asyncapi: null,
         error: err,
       };
     }
-
-    return { data: null, asyncapi: null, error: err };
-  };
-
-  private extractDocument = (data: any): ParserReturn => {
-    if (data.json instanceof Function) {
-      return {
-        data: data.json(),
-        asyncapi: data,
-      };
-    }
-    if (typeof data._json === 'object') {
-      return {
-        data: data._json,
-        asyncapi: data,
-      };
-    }
-    return {
-      data,
-      asyncapi: data,
-    };
+    return { error: err };
   };
 }
+
+export const parser = new Parser();
