+++ library/src/index.ts
@@ -1,8 +1,9 @@
 import AsyncApi from './containers/AsyncApi/AsyncApi';
+
+export { AsyncApiProps } from './containers/AsyncApi/AsyncApi';
 export { ConfigInterface } from './config/config';
 export {
   AsyncAPI as AsyncApiInterface,
-  AsyncApiProps,
   FetchingSchemaInterface,
 } from './types';
